refactor(about): split C++ into multiple files
Everything seems to work, and I'm no longer rapidly iterating.
